requirements.txt
pyautogui
pygame
pytesseract
pillow